Cleanup database initialization of `PpdbSqlBase` sub-classes
This moves the `init_database` methods to sub-classes and removes it
from the base class so that it is not polymorphic. The BigQuery init
method was cleaned up to use its config class.
